+++ mod/class.tx_rnbase_mod_Tables.php
@@ -216,6 +216,13 @@ class tx_rnbase_mod_Tables {
 		return $overlays;
 	}
 
+	/**
+	 *
+	 * @param tx_rnbase_model_data $options
+	 * @param tx_rnbase_model_base $obj
+	 * @param tx_rnbase_util_FormTool $formTool
+	 * @return string
+	 */
 	private static function addLinker($options, $obj, $formTool) {
 		$out = '';
 		$linkerArr = $options->getLinker();
@@ -223,6 +230,17 @@ class tx_rnbase_mod_Tables {
 			$linkerimplode = $options->getLinkerimplode() ? $options->getLinkerimplode() : '<br />';
 			$currentPid = (int) $options->getPid();
 			foreach($linkerArr As $linker) {
+				if (!$linker instanceof tx_rnbase_mod_LinkerInterface) {
+					// backward compatibility, the interface with the makeLink method is new!
+					if (!is_callable(array($linker, 'makeLink'))) {
+						throw new Exception(
+							'Linker "' . get_class($linker) . '" has to implement interface "tx_rnbase_mod_LinkerInterface".'
+						);
+					}
+					t3lib_div::deprecationLog(
+						'Linker "' . get_class($linker) . '" has to implement interface "tx_rnbase_mod_LinkerInterface".'
+					);
+				}
 				$out .= $linker->makeLink($obj, $formTool, $currentPid, $options);
 				$out .= $linkerimplode;
 			}